requirements.txt
fastapi
uvicorn
python-dotenv
psycopg[binary]
pydantic